FlyControls and fixed OrbitControls
